Renamed Static to Compiler
